moves Fatal from Storage into Amplify Core and put preconditionFailure into Fatal
