+++ AmplifyPlugins/DataStore/AWSDataStoreCategoryPlugin/Storage/SQLite/SQLStatement+CreateTable.swift
@@ -51,7 +51,7 @@ struct CreateTableStatement: SQLStatement {
             statement += "  foreign key(\"\(foreignKey.sqlName)\") "
             let associatedModel = foreignKey.requiredAssociatedModelName
             guard let schema = ModelRegistry.modelSchema(from: associatedModel) else {
-                return Amplify.preconditionFailure("""
+                return Fatal.preconditionFailure("""
                 Could not retrieve schema for the model \(associatedModel), verify that datastore is initialized.
                 """)
             }
