+++ AmplifyPlugins/DataStore/AWSDataStoreCategoryPlugin/Storage/SQLite/StorageEngineAdapter+SQLite.swift
@@ -83,7 +83,7 @@ final class SQLiteStorageEngineAdapter: StorageEngineAdapter {
 
     static func getDbFilePath(databaseName: String) -> URL {
         guard let documentsPath = getDocumentPath() else {
-            return Amplify.preconditionFailure("Could not create the database. The `.documentDirectory` is invalid")
+            return Fatal.preconditionFailure("Could not create the database. The `.documentDirectory` is invalid")
         }
         return documentsPath.appendingPathComponent("\(databaseName).db")
     }
